+++ libnavui-tripprogress/src/main/java/com/mapbox/navigation/ui/tripprogress/TripProgressResult.kt
@@ -8,4 +8,25 @@ internal sealed class TripProgressResult {
         val totalTimeRemaining: Double,
         val percentRouteTraveled: Double
     ) : TripProgressResult()
+
+    sealed class TripOverview : TripProgressResult() {
+        data class RouteLegTripOverview(
+            val legIndex: Int,
+            val legTime: Double,
+            val legDistance: Double,
+            val estimatedTimeToArrival: Long,
+        )
+
+        data class Success(
+            val routeLegTripDetail: List<RouteLegTripOverview>,
+            val totalTime: Double,
+            val totalDistance: Double,
+            val totalEstimatedTimeToArrival: Long,
+        ) : TripOverview()
+
+        data class Failure(
+            val errorMessage: String?,
+            val throwable: Throwable?
+        ) : TripOverview()
+    }
 }